Updated triangle intersection code to use fp32 instead of fp64.

+++ src/triangle.cpp
@@ -7,33 +7,33 @@ bool IntersectTriangle(const Ray& ray, const Triangle& triangle,
   Vector edge2 = triangle.points[2] - triangle.points[0];
 
   Vector p = CrossProduct(ray.GetDirection(), edge2);
-  double divisor = DotProduct(edge1, p);
+  float divisor = DotProduct(edge1, p);
 
   // todo: do we need to check against epsilon for better numeric stability?
   if (divisor == 0.0)
     return false;
 
-  const double invDivisor = 1.0 / divisor;
+  const float invDivisor = 1.0f / divisor;
 
   // compute barycentric coordinate b1
   Vector t = ray.GetOrigin() - triangle.points[0];
-  double b1 = invDivisor * DotProduct(t, p);
+  float b1 = invDivisor * DotProduct(t, p);
   if (b1 < 0.0 || b1 > 1.0)
     return false;
 
   // compute barycentric coordnate b2
   Vector q = CrossProduct(t, edge1);
-  double b2 = invDivisor * DotProduct(ray.GetDirection(), q);
+  float b2 = invDivisor * DotProduct(ray.GetDirection(), q);
   if (b2 < 0.0 || b1 + b2 > 1.0)
     return false;
 
   // compute distance from ray origin to intersection point
-  double distance = invDivisor * DotProduct(edge2, q);
+  float distance = invDivisor * DotProduct(edge2, q);
   if (distance < 0.0)
     return false;
 
   intersection.t = distance;
-  intersection.epsilon = 1e-3 * distance;
+  intersection.epsilon = 1e-3f * distance;
   intersection.b1 = b1;
   intersection.b2 = b2;
   return true;
